% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFinYield.R
\name{calcFinYield}
\alias{calcFinYield}
\title{Calc financial yield}
\usage{
calcFinYield(om, f1 = 100, f2 = 0.15, f3 = 170, cfg = loadConfig())
}
\arguments{
\item{om}{outputsModel}

\item{f1}{Coût fixe (€/ha)}

\item{f2}{Coût de l'eau et de l'énergie par m³ d'eau appliquée (€/m³)}

\item{f3}{Court du maïs (€/t)}

\item{cfg}{a config object. Configuration to use. See \link{loadConfig} for details}
}
\value{
outputsModel
}
\description{
Calc financial yield
}
