% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimateHorizonProjections.R
\name{getClimateHorizonProjections}
\alias{getClimateHorizonProjections}
\title{Get the climate horizon projections}
\usage{
getClimateHorizonProjections(datasets, scenario, step, vars, i, outputsModel)
}
\arguments{
\item{datasets}{\link{data.frame} The datasets containing the scenario and dates
information.}

\item{scenario}{\link{character} The scenario for which the climate horizon
projections are requested.}

\item{step}{\link{character} The time step for which the climate horizon
projections are requested.}

\item{vars}{\link{character} The variables for which the climate horizon
projections are requested.}

\item{outputsModel}{\link{list} See \code{\link{optirrigCORE::getOutputsModel}}.}
}
\description{
This function returns the climate horizon projections for a given scenario,
time step, and variables.
}
