% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPossibleDecisions.R
\name{getPossibleDecisions}
\alias{getPossibleDecisions}
\title{Get the possible water supply decisions}
\usage{
getPossibleDecisions(step = "t1")
}
\arguments{
\item{step}{\link{character} The time step for which the possible decisions are
requested.}
}
\value{
\link{numeric} The possible water supply decisions.
}
\description{
This function returns the possible water supply decisions for a given time
step.
}
