% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\docType{package}
\name{OptirrigSIA-package}
\alias{OptirrigSIA}
\alias{OptirrigSIA-package}
\title{OptirrigSIA: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kevin Bosirany ORLANDO \email{kevinbosirany@gmail.com} (\href{https://orcid.org/0009-0009-2784-3108}{ORCID})

}
\keyword{internal}
