% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlots.R
\name{savePlots}
\alias{savePlots}
\title{Save plots}
\usage{
savePlots(
  scenarios_final,
  i,
  outputsModel,
  fun_prod,
  version,
  force = FALSE,
  cfg = loadConfig()
)
}
\arguments{
\item{scenarios_final}{\link{data.frame} Scenarios}

\item{i}{\link{integer} Line index of scenarios_final}

\item{outputsModel}{See return of \code{\link{loadOutputsModel}}}

\item{version}{\link{character} Version of the dataset}

\item{cfg}{a config object. Configuration to use. See \link{loadConfig} for details}
}
\description{
Save plots for each scenario
}
